requirements.txt
cython==3.0.12
colorama==0.4.6
requests==2.32.3
requests[sock]==2.32.3
winshell==0.6
pycryptodome==3.21.0
discord-webhook==1.3.1
pywin32==308
setuptools==75.8.0
flask==3.1.0
